group http and grpc modules
