+++ settings.gradle
@@ -20,10 +20,8 @@ include 'tracing-bom'
 include 'tracing-core'
 include 'tracing-jaeger'
 include 'tracing-opentelemetry'
-include 'tracing-opentelemetry-http-client'
-include 'tracing-opentelemetry-http-server'
-include 'tracing-opentelemetry-grpc-client'
-include 'tracing-opentelemetry-grpc-server'
+include 'tracing-opentelemetry-http'
+include 'tracing-opentelemetry-grpc'
 include 'tracing-zipkin'
 include 'tests:kotlin-tests'
