--- tracing-opentelemetry-http-client/src/main/java/io.micronaut.tracing.opentelemetry.instrument.http/HttpRequestSetter.java
@@ -1,28 +1,0 @@
-/*
- * Copyright 2017-2022 original authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.micronaut.tracing.opentelemetry.instrument.http;
-
-import io.micronaut.http.MutableHttpRequest;
-import io.opentelemetry.context.propagation.TextMapSetter;
-
-enum HttpRequestSetter implements TextMapSetter<MutableHttpRequest> {
-    INSTANCE;
-
-    @Override
-    public void set(MutableHttpRequest httpRequest, String key, String value) {
-        httpRequest.header(key, value);
-    }
-}
