--- tracing-opentelemetry-http-server/build.gradle
@@ -1,24 +1,0 @@
-plugins {
-    id 'io.micronaut.build.internal.tracing-module'
-    id 'com.google.protobuf'
-}
-
-dependencies {
-    api mn.micronaut.runtime
-    api mn.micronaut.http.client
-    api projects.tracingAnnotation
-    api projects.tracingOpentelemetry
-    api libs.opentelemetry.api
-    api libs.opentelemetry.extension.annotations
-    api libs.opentelemetry.instrumentation.api
-    api libs.opentelemetry.instrumentation.api.semconv
-
-    compileOnly libs.kotlinx.coroutines.core
-    compileOnly libs.kotlinx.coroutines.reactor
-
-    testImplementation mn.reactor
-    testImplementation mn.micronaut.reactor.http.client
-
-    testImplementation libs.opentelemetry.sdk
-    testImplementation libs.opentelemetry.sdk.testing
-}
