--- tracing-opentelemetry-http-server/src/main/java/io.micronaut.tracing.opentelemetry.instrument.http/HttpRequestGetter.java
@@ -1,38 +1,0 @@
-/*
- * Copyright 2017-2022 original authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.micronaut.tracing.opentelemetry.instrument.http;
-
-import io.micronaut.core.annotation.Nullable;
-import io.micronaut.http.HttpRequest;
-import io.opentelemetry.context.propagation.TextMapGetter;
-
-enum HttpRequestGetter implements TextMapGetter<HttpRequest> {
-    INSTANCE;
-
-    @Override
-    public Iterable<String> keys(HttpRequest request) {
-        return request.getHeaders().names();
-    }
-
-    @Override
-    @Nullable
-    public String get(@Nullable HttpRequest request, String key) {
-        if (request == null) {
-            return null;
-        }
-        return request.getHeaders().getFirst(key).orElse(null);
-    }
-}
