--- tracing-opentelemetry-http-server/src/test/groovy/io/micronaut/tracing/instrument/IExecutorService.groovy
@@ -1,7 +1,0 @@
-package io.micronaut.tracing.instrument
-
-import java.util.concurrent.ExecutorService
-
-interface IExecutorService extends ExecutorService {
-    void newMethod()
-}
