+++ tracing-opentelemetry-http/src/main/java/io/micronaut/tracing/opentelemetry/instrument/http/client/MicronautHttpNetClientAttributesGetter.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2017-2022 original authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.micronaut.tracing.opentelemetry.instrument.http.client;
+
+import io.micronaut.http.HttpResponse;
+import io.micronaut.http.MutableHttpRequest;
+import io.opentelemetry.instrumentation.api.instrumenter.net.InetSocketAddressNetClientAttributesGetter;
+import io.opentelemetry.semconv.trace.attributes.SemanticAttributes;
+
+import javax.annotation.Nullable;
+import java.net.InetSocketAddress;
+
+final class MicronautHttpNetClientAttributesGetter
+    extends InetSocketAddressNetClientAttributesGetter<MutableHttpRequest, HttpResponse> {
+
+    @Override
+    public InetSocketAddress getAddress(MutableHttpRequest request, @Nullable HttpResponse response) {
+        return request.getRemoteAddress();
+    }
+
+    @Override
+    public String transport(MutableHttpRequest request, @Nullable HttpResponse response) {
+        return SemanticAttributes.NetTransportValues.IP_TCP;
+    }
+
+}
